requirements.txt
flask==2.3.3
Flask-RESTful==0.3.10
python-docx==0.8.11
psycopg2-binary==2.9.6
sentence-transformers==2.2.2
transformers==4.30.2
huggingface_hub==0.14.1
datasketch==1.5.1
dask==2023.5.0
nltk==3.8.1
pymorphy2==0.9.1
sqlalchemy==2.0.19
pgvector==0.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.25.0
scikit-learn==1.3.0
torch==2.0.1
requests==2.31.0 